requirements.txt
streamlit==1.34.0
langchain==0.1.17
nltk==3.8.1
langchain-google-genai==0.1.1
langchain-huggingface==0.1.2
faiss-cpu==1.7.4
unstructured==0.11.8
python-dotenv==1.0.1
streamlit-chat
# Pre-download 'punkt' and 'averaged_perceptron_tagger' using nltk.download()